backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.9.0
attrs==25.3.0
azure-cognitiveservices-speech==1.44.0
azure-core==1.34.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
crypto==1.4.1
cryptography==45.0.4
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
elevenlabs==1.59.0
email_validator==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.13
fastapi-cli==0.0.7
fastapi-docx==1.0.3
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.5.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.0
idna==3.10
importlib_metadata==8.6.1
janus==1.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-google-community==2.0.7
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.42
loguru==0.7.3
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
miniaudio==1.61
motor==3.7.1
multidict==6.4.4
mypy_extensions==1.1.0
Naked==0.1.32
ngrok==1.4.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.82.0
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
phonenumbers==8.13.55
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.13.2
pyparsing==3.2.3
pypdfium2==4.30.1
python-docx==1.2.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.6
rpds-py==0.25.1
rsa==4.9.1
sentry-sdk==2.29.1
shellescape==3.8.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.4.7
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.7.0
tokenizers==0.21.1
tqdm==4.67.1
twilio==9.6.1
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.0.5
websockets==12.0
win32_setctime==1.2.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
